+++ app/models/journey.rb
@@ -1,4 +1,6 @@
 class Journey < ApplicationRecord
   self.implicit_order_column = "created_at"
   has_many :steps
+
+  validates :liquid_template, presence: true
 end