+++ app/models/step.rb
@@ -8,6 +8,8 @@ class Step < ApplicationRecord
   has_one :single_date_answer
   has_one :checkbox_answers
 
+  scope :that_are_questions, -> { where(contentful_model: "question") }
+
   def answer
     @answer ||=
       radio_answer ||