+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog 1.0.0].
 
 ## [Unreleased]
 
+- users can see an initial slice of their specification as HTML at the end of their journey
+
 ## [release-004] - 2020-12-17
 
 - fix primary key type on long_text_answers table to UUID